Beyond the CRUD: Architecting E-Commerce Systems That Survive Real-World Threats_task_3/requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6